achievements 5: debug row, docs, and score only the save in memory
The debug page gets a seventh row (grid refitted to 20px buttons at a
22px pitch): TEST queues a toast, RESYNC re-derives this save's
achievements after a second tap, and the note counts any text too wide
for the TROPHY list.

Evaluation now also requires the save block's trainer ID to match the
adopted playthrough. Without it a soft reset after an unsaved new game
would score the reloaded old save under the new game's record.
AchView's counters are u32, so the 100,000-step goal no longer
truncates.

README, README-TECHNICAL, the cheatsheet (third overlay, sixth tab,
top[8]/top[9], the y 0/40/152 split, two pitfalls), ROADMAP Part E and
the plan's "how it landed" notes.

+++ 3ds/achievements.h
@@ -29,8 +29,8 @@ struct AchView
 {
     const char *title;
     const char *desc;
-    u16   progress;   // how far along, clamped to goal
-    u16   goal;       // above 1: a counter the UI may draw while locked
+    u32   progress;   // how far along, clamped to goal
+    u32   goal;       // above 1: a counter the UI may draw while locked
     bool8 unlocked;
     bool8 unseen;     // unlocked and not yet shown on the TROPHY tab
 };
